+++ app/src/main/java/com/checkmyself/onboarding/ui/activity/RegisterActivity.kt
@@ -146,8 +146,8 @@ class RegisterActivity : BaseActivity(), View.OnClickListener {
                         binding.emailEdit.text.toString().trim(),
                         dob,
                         gender,
-                        binding.weightEdit.text.toString().trim().toInt(),
-                        binding.heightEdit.text.toString().trim().toInt(),
+                        binding.weightEdit.text.toString().trim().toDouble(),
+                        binding.heightEdit.text.toString().trim().toDouble(),
                         null
                     )
                     CreatePasswordActivity.startActivity(this, request)
